ABDMP-355 - added SurfaceData repository, removed floats

+++ Abdm.Calculation.WebApi/RequestModels/SurfaceRequestModel.cs
@@ -71,10 +71,10 @@
 
     public class SurfaceDataItemRequestModel
     {
-        public float x { get; set; }
+        public double x { get; set; }
 
-        public float y { get; set; }
+        public double y { get; set; }
 
-        public float z { get; set; }
+        public double z { get; set; }
     }
 }
